Moving setup_heartbeat to session.py
